US138 T140 Refactors GameManagerStateTests
